+++ dyninstAPI/src/codegen-aarch64.C
@@ -91,11 +91,11 @@ void insnCodeGen::generateTrap(codeGen &gen) {
 void insnCodeGen::generateBranch(codeGen &gen, long disp, bool link) {
     if (labs(disp) > MAX_BRANCH_OFFSET) {
         fprintf(stderr, "ABS OFF: 0x%lx, MAX: 0x%lx\n",
-                labs(disp), (unsigned long) MAX_BRANCH_OFFSET);
-        bperr( "Error: attempted a branch of 0x%lx\n", disp);
+                (unsigned long)labs(disp), (unsigned long) MAX_BRANCH_OFFSET);
+        bperr( "Error: attempted a branch of 0x%lx\n", (unsigned long)disp);
         logLine("a branch too far\n");
         showErrorCallback(52, "Internal error: branch too far");
-        bperr( "Attempted to make a branch of offset 0x%lx\n", disp);
+        bperr( "Attempted to make a branch of offset 0x%lx\n", (unsigned long)disp);
         assert(0);
     }
 
@@ -201,11 +201,11 @@ void insnCodeGen::generateBranchViaTrap(codeGen &gen, Address from, Address to, 
     } else {
         // Too far to branch and no proc to register trap.
         fprintf(stderr, "ABS OFF: 0x%lx, MAX: 0x%lx\n",
-                labs(disp), (unsigned long) MAX_BRANCH_OFFSET);
-        bperr( "Error: attempted a branch of 0x%lx\n", disp);
+                (unsigned long)labs(disp), (unsigned long) MAX_BRANCH_OFFSET);
+        bperr( "Error: attempted a branch of 0x%lx\n", (unsigned long)disp);
         logLine("a branch too far\n");
         showErrorCallback(52, "Internal error: branch too far");
-        bperr( "Attempted to make a branch of offset 0x%lx\n", disp);
+        bperr( "Attempted to make a branch of offset 0x%lx\n", (unsigned long)disp);
         assert(0);
     }
 }
